decomposition to custom useGame hook
